# specs/001-company-role-user-setup/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GenzFeast — Company, Role, Category, Department & User Creation API
  version: "1.0.0"
  description: >
    Contract for feature 001-company-role-user-setup. All endpoints require
    Authorization: Bearer <JWT> (Architecture §4). company_id/role are always
    taken from the verified JWT claims, never from the request body or path.
    Every write additionally lands in audit_logs automatically via DB trigger
    (out of band for this contract).
servers:
  - url: https://api.genzfeast.example/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/users:
    get:
      tags: [Users]
      summary: List every user across every company (System Admin only)
      responses:
        "200":
          description: All users, platform-wide
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: >
        Create a Company Admin or Staff user for any company (System Admin
        only). Creating a Student isn't supported here — that stays
        self-registration only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTenantUserCreateRequest" }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
  /admin/users/{userId}:
    patch:
      tags: [Users]
      summary: >
        Edit a Company Admin or Staff user, any company (System Admin only).
        PATCH semantics — only submitted fields change; omitting `password`
        keeps the existing one. Same role restriction as creation: the
        existing account, and any submitted new role, must be
        company_staff/company_admin.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTenantUserUpdateRequest" }
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
    delete:
      tags: [Users]
      summary: >
        Soft-delete a user, any role, any company (System Admin only). Blocks
        that user from logging in ("blocked contact admin") and immediately
        revokes their active sessions — never a hard delete. A System Admin
        cannot delete their own account this way.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: User soft-deleted; their sessions revoked }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/departments/{departmentId}:
    patch:
      tags: [Departments]
      summary: Update a department belonging to any company (System Admin only)
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataUpdateRequest" }
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Departments]
      summary: Soft-delete a department belonging to any company (System Admin only)
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      responses:
        "204": { description: Department soft-deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /tenant/departments:
    get:
      tags: [Departments]
      summary: >
        List departments for the caller's own company (Company Admin, Staff,
        Student). Server-side paginated (revised — this task's own explicit
        ask); `page`/`limit` both default when omitted, so an existing caller
        needing the full list (e.g. a Student's profile dropdown) can still
        pass a high `limit` rather than paging through results.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: A page of the caller's company departments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedDepartmentsResponse" }
    post:
      tags: [Departments]
      summary: Create a department scoped to the caller's own company (Company Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataCreateRequest" }
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/departments/{departmentId}:
    get:
      tags: [Departments]
      summary: >
        Not in this feature's original scope — added once `list` became
        paginated, so the edit screen can look up one department directly
        rather than fetch-the-full-list-and-find-by-id (Company Admin,
        Staff, Student).
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      responses:
        "200":
          description: The department
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Departments]
      summary: Update a department scoped to the caller's own company (Company Admin only)
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataUpdateRequest" }
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Departments]
      summary: Soft-delete a department scoped to the caller's own company (Company Admin only)
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      responses:
        "204": { description: Department soft-deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tenant/users:
    get:
      tags: [Users]
      summary: >
        List company_staff/student/teaching/non_teaching users for the
        caller's own company (Company Admin only). Revised — company_admin
        dropped from this screen's manageable-role set on explicit request.
      responses:
        "200":
          description: The caller's company users (manageable roles only)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: Create a company_staff/student/teaching/non_teaching user in the caller's own company (Company Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantUserCreateRequest" }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
  /tenant/users/{userId}:
    patch:
      tags: [Users]
      summary: >
        Not in this feature's original scope — this task's own general edit,
        mirroring System Admin's /admin/users/{userId} PATCH but scoped to
        the caller's own company and TENANT_MANAGEABLE_ROLES.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantUserUpdateRequest" }
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
    delete:
      tags: [Users]
      summary: >
        Not in this feature's original scope — this task's own soft delete,
        mirroring System Admin's /admin/users/{userId} DELETE but scoped to
        the caller's own company. Blocks the user from logging in ("blocked
        contact admin") and immediately revokes their active sessions. A
        Company Admin cannot delete their own account this way.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: User soft-deleted; their sessions revoked }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tenant/users/{userId}/status:
    patch:
      tags: [Users]
      summary: Activate or deactivate a Staff/Student user in the caller's own company (Company Admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, inactive] }
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tenant/users/{userId}/role:
    patch:
      tags: [Users]
      summary: >
        Not in this feature's original scope — added alongside the Company
        Admin management screens for an explicit "change user role"
        capability. Reassigns among TENANT_MANAGEABLE_ROLES only (same
        restriction as creation); never touches a Company Admin or System
        Admin account (Company Admin only).
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [company_staff, student, teaching, non_teaching] }
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DepartmentId:
      name: departmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
    SearchParam:
      name: search
      in: query
      required: false
      description: Case-insensitive name substring filter
      schema: { type: string }
  responses:
    Forbidden:
      description: Caller's role/company scope does not permit this action
    NotFound:
      description: Resource does not exist or is not visible to the caller's scope
    ValidationError:
      description: Request failed field-level validation
    DuplicateUsername:
      description: Username already exists within the target company
  schemas:
...
    Department:
      type: object
      properties:
        id: { type: string, format: uuid }
        company_id: { type: string, format: uuid }
        name: { type: string, example: "Computer Science" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaginatedDepartmentsResponse:
      type: object
      description: Server-side pagination (10 per page) — the mobile list screen fetches the next page on scroll.
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Department" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    MasterDataCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    MasterDataUpdateRequest:
      type: object
      properties:
        name: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        company_id: { type: string, format: uuid, nullable: true }
        role: { type: string, enum: [system_admin, company_admin, company_staff, student, teaching, non_teaching] }
        name: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender], nullable: true }
        category_id: { type: string, format: uuid, nullable: true }
        department_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [active, inactive, locked] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    TenantUserCreateRequest:
      type: object
      required: [name, username, password, email, gender, role]
      description: >
        `role` revised (company_admin dropped from this screen on explicit
        request; student/teaching/non_teaching added). `category_id` is
        required exactly when role is 'student' (mirrors
        StudentRegisterRequest's own requirement); `department_id` stays
        optional and is ignored for every other role.
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password, description: "Temporary password" }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender] }
        role: { type: string, enum: [company_staff, student, teaching, non_teaching] }
        category_id: { type: string, format: uuid, description: "Required when role is 'student'" }
        department_id: { type: string, format: uuid }
    TenantUserUpdateRequest:
      type: object
      description: PATCH semantics — every field optional, only submitted fields change. Omitting `password` keeps the existing one.
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender] }
        role: { type: string, enum: [company_staff, student, teaching, non_teaching] }
        category_id: { type: string, format: uuid }
        department_id: { type: string, format: uuid }

    # System Admin's cross-tenant equivalent of TenantUserCreateRequest —
    # same shape plus an explicit company_id, since the caller isn't scoped
    # to one Company. Same role restriction (no Student here).
    AdminTenantUserCreateRequest:
      type: object
      required: [name, username, password, email, gender, role, company_id]
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password, description: "Temporary password" }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender] }
        role: { type: string, enum: [company_staff, company_admin] }
        company_id: { type: string, format: uuid }
    AdminTenantUserUpdateRequest:
      type: object
      description: PATCH semantics — every field optional, only submitted fields change.
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password, description: "Omit to keep the existing password" }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender] }
        role: { type: string, enum: [company_staff, company_admin] }
        company_id: { type: string, format: uuid }
...